LayoutTests/imported/w3c/web-platform-tests/svg/linking/scripted/a.ping-functionality-expected.txt


FAIL SVG anchor with ping attribute should send ping on click promise_test: Unhandled rejection with value: object "ReferenceError: Can't find variable: observe_entry"
FAIL SVG anchor with multiple ping URLs should send multiple pings promise_test: Unhandled rejection with value: object "ReferenceError: Can't find variable: observe_entry"
FAIL SVG anchor ping attribute should be settable via ping IDL attribute assert_equals: expected (string) "https://example.com/ping1 https://example.com/ping2" but got (undefined) undefined
